Phase 4: in-process command tests; E2E suite runs the NativeAOT binary in CI
The 86 subprocess integration tests are restructured:

- 72 tests port to the new in-process CommandTests: each invocation
  builds a fresh CommandContext (FakeConsole, TestKeyYubiKeyService,
  temp config dir) and dispatches through CliRunner — the same
  dispatch/error-mapping path the real binary uses. Test bodies are
  unchanged from the subprocess versions; only the harness differs.
  The whole in-process suite runs in ~2 s with no process spawning.
- 15 tests that genuinely need a real process stay in EndToEndTests
  (renamed from ProgramTests, tagged [Trait(Category, E2E)]): run
  subprocess execution and PTY redaction, Ctrl+C via PTY, installscript
  (Environment.ProcessPath), and composition-root env-var handling.

Supporting changes:
- CliRunner extracts the dispatch + exception-to-exit-code mapping from
  Program.cs so tests and the binary share it exactly
- IConsole gains In (stdin reader); ingest reads via the seam, making
  it scriptable in-process. FakeConsole added to TswapTests
- TswapBinaryFixture honours TSWAP_E2E_BINARY, letting CI point the E2E
  suite at a published binary. CI now runs E2E against a Debug NativeAOT
  publish on Linux — the first time the AOT-compiled binary itself is
  tested (Debug because the TSWAP_TEST_KEY bypass is Debug-only)
- New unit tests for logic that had no direct coverage: VaultUnlocker
  (XOR either-key reconstruction, legacy challenge, serial selection,
  override key) and CliEnvironment (config-dir resolution branches,
  legacy migration, prefix detection)
- runtests.sh: --unit (~5 s) / --e2e; CI unit step filters Category!=E2E
  on all three OSes

TswapTests now references TswapCli (ValidateExecutableReferencesMatch-
SelfContained=false — RID-specific exe referenced by RID-agnostic test
project). 314 tests pass: 285 TswapTests + 14 ConsoleIntercept.Tests
in-process, 15 E2E. NativeAOT Release publish verified.
