--- TswapTests/ProgramTests.cs
@@ -1,1726 +1,0 @@
-using System.Diagnostics;
-using System.Runtime.InteropServices;
-using System.Security.Cryptography;
-using System.Text;
-using System.Text.Json;
-using TswapCore;
-using Xunit;
-
-namespace TswapTests;
-
-/// <summary>
-/// Integration tests for Program.cs using TSWAP_TEST_KEY to bypass YubiKey.
-/// Each test gets an isolated temp config directory and a deterministic test key.
-/// The tswap binary is built once per run by <see cref="TswapBinaryFixture"/> and
-/// invoked directly (not via `dotnet run`, which re-evaluates MSBuild per call).
-/// </summary>
-public class ProgramTests : IClassFixture<TswapBinaryFixture>, IDisposable
-{
-    private readonly string _tempDir;
-    private readonly string _testKeyHex;
-    private readonly string _binaryPath;
-
-    public ProgramTests(TswapBinaryFixture binary)
-    {
-        _tempDir = Path.Combine(Path.GetTempPath(), "tswap-prog-test-" + Guid.NewGuid().ToString("N")[..8]);
-        Directory.CreateDirectory(_tempDir);
-
-        // Deterministic 32-byte test key
-        _testKeyHex = "0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF0123456789ABCDEF";
-        _binaryPath = binary.BinaryPath;
-    }
-
-    public void Dispose()
-    {
-        if (Directory.Exists(_tempDir))
-            Directory.Delete(_tempDir, true);
-    }
-
-    private ProcessStartInfo MakePsi(bool redirectStdin = false)
-    {
-        var psi = new ProcessStartInfo
-        {
-            FileName = _binaryPath,
-            RedirectStandardOutput = true,
-            RedirectStandardError = true,
-            RedirectStandardInput = redirectStdin,
-            UseShellExecute = false,
-            CreateNoWindow = true
-        };
-        psi.Environment["TSWAP_TEST_KEY"] = _testKeyHex;
-        psi.Environment["TSWAP_TEST_SUDO_BYPASS"] = "1";
-        psi.Environment["TSWAP_CONFIG_DIR"] = _tempDir;
-        // fork() from a JIT-mode .NET process crashes when W^X protection is active.
-        // The deployed AOT binary has no JIT so this is not needed in production.
-        psi.Environment["DOTNET_EnableWriteXorExecute"] = "0";
-        return psi;
-    }
-
-    private (int exitCode, string stdout, string stderr) RunTswap(params string[] args)
-    {
-        var psi = MakePsi();
-        foreach (var arg in args)
-            psi.ArgumentList.Add(arg);
-
-        using var process = Process.Start(psi)!;
-        var stdout = process.StandardOutput.ReadToEnd();
-        var stderr = process.StandardError.ReadToEnd();
-        process.WaitForExit();
-
-        return (process.ExitCode, stdout, stderr);
-    }
-
-    private (int exitCode, string stdout, string stderr) RunTswapWithStdin(string stdin, params string[] args)
-    {
-        var psi = MakePsi(redirectStdin: true);
-        foreach (var arg in args)
-            psi.ArgumentList.Add(arg);
-
-        using var process = Process.Start(psi)!;
-        process.StandardInput.Write(stdin);
-        process.StandardInput.Close();
-        var stdout = process.StandardOutput.ReadToEnd();
-        var stderr = process.StandardError.ReadToEnd();
-        process.WaitForExit();
-
-        return (process.ExitCode, stdout, stderr);
-    }
-
-    // --- Init ---
-
-    [Fact]
-    public void Init_CreatesConfigFile()
-    {
-        var (exit, stdout, _) = RunTswap("init");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("test mode", stdout);
-        Assert.True(File.Exists(Path.Combine(_tempDir, "config.json")));
-    }
-
-    [Fact]
-    public void Init_ConfigHasSyntheticSerials()
-    {
-        RunTswap("init");
-
-        var json = File.ReadAllText(Path.Combine(_tempDir, "config.json"));
-        var config = JsonSerializer.Deserialize(json, TswapJsonContext.Default.Config)!;
-
-        Assert.Equal(new List<int> { 99999999, 99999998 }, config.YubiKeySerials);
-    }
-
-    // --- Create ---
-
-    [Fact]
-    public void Create_GeneratesSecret()
-    {
-        RunTswap("init");
-        var (exit, stdout, _) = RunTswap("create", "my-secret");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("my-secret", stdout);
-        Assert.Contains("created", stdout);
-    }
-
-    [Fact]
-    public void Create_CustomLength()
-    {
-        RunTswap("init");
-        var (exit, stdout, _) = RunTswap("create", "short-secret", "8");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("8 chars", stdout);
-    }
-
-    [Fact]
-    public void Create_DuplicateNameFails()
-    {
-        RunTswap("init");
-        RunTswap("create", "dup-secret");
-        var (exit, _, stderr) = RunTswap("create", "dup-secret");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("already exists", stderr);
-    }
-
-    [Fact]
-    public void Create_EmptyNameFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("empty", stderr);
-    }
-
-    [Fact]
-    public void Create_NameWithSpaceFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "bad name");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("Invalid secret name", stderr);
-    }
-
-    [Fact]
-    public void Create_NameWithSpecialCharsFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "bad!name");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("Invalid secret name", stderr);
-    }
-
-    [Fact]
-    public void Create_NameTooLongFails()
-    {
-        RunTswap("init");
-        var longName = new string('a', 65);
-        var (exit, _, stderr) = RunTswap("create", longName);
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("too long", stderr);
-    }
-
-    [Fact]
-    public void Create_NameAtMaxLengthSucceeds()
-    {
-        RunTswap("init");
-        var maxName = new string('a', 64);
-        var (exit, _, _) = RunTswap("create", maxName);
-
-        Assert.Equal(0, exit);
-    }
-
-    [Fact]
-    public void Create_ZeroLengthFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "valid-name", "0");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("at least 1", stderr);
-    }
-
-    [Fact]
-    public void Create_NegativeLengthFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "valid-name", "-5");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("at least 1", stderr);
-    }
-
-    [Fact]
-    public void Create_TooLongLengthFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "valid-name", "4097");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("at most 4096", stderr);
-    }
-
-    [Fact]
-    public void Create_NonNumericLengthFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("create", "valid-name", "abc");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("whole number", stderr);
-    }
-
-    [Fact]
-    public void Create_MaxLengthSucceeds()
-    {
-        RunTswap("init");
-        var (exit, stdout, _) = RunTswap("create", "valid-name", "4096");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("4096 chars", stdout);
-    }
-
-    // --- Names ---
-
-    [Fact]
-    public void Names_ListsCreatedSecrets()
-    {
-        RunTswap("init");
-        RunTswap("create", "alpha");
-        RunTswap("create", "beta");
-
-        var (exit, stdout, _) = RunTswap("names");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("alpha", stdout);
-        Assert.Contains("beta", stdout);
-    }
-
-    [Fact]
-    public void Names_EmptyVault()
-    {
-        RunTswap("init");
-        var (exit, stdout, _) = RunTswap("names");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("No secrets", stdout);
-    }
-
-    // --- Ingest ---
-
-    [Fact]
-    public void Ingest_StoresStdinValue()
-    {
-        RunTswap("init");
-        var (exit, stdout, _) = RunTswapWithStdin("piped-secret-value", "ingest", "from-stdin");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("ingested", stdout);
-
-        // Verify it appears in names
-        var (_, namesOut, _) = RunTswap("names");
-        Assert.Contains("from-stdin", namesOut);
-    }
-
-    [Fact]
-    public void Ingest_DuplicateFails()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("value1", "ingest", "dup-ingest");
-        var (exit, _, stderr) = RunTswapWithStdin("value2", "ingest", "dup-ingest");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("already exists", stderr);
-    }
-
-    [Fact]
-    public void Ingest_EmptyNameFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswapWithStdin("some-value", "ingest", "");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("empty", stderr);
-    }
-
-    [Fact]
-    public void Ingest_NameWithSpaceFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswapWithStdin("some-value", "ingest", "bad name");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("Invalid secret name", stderr);
-    }
-
-    [Fact]
-    public void Ingest_TooLongValueFails()
-    {
-        RunTswap("init");
-        var longValue = new string('x', 65537);
-        var (exit, _, stderr) = RunTswapWithStdin(longValue, "ingest", "toolong");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("too long", stderr);
-    }
-
-    [Fact]
-    public void Ingest_MaxLengthValueSucceeds()
-    {
-        RunTswap("init");
-        var maxValue = new string('x', 65536);
-        var (exit, stdout, _) = RunTswapWithStdin(maxValue, "ingest", "maxlen");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("ingested", stdout);
-    }
-
-    // --- Burn ---
-
-    [Fact]
-    public void Burn_MarksSecretAsBurned()
-    {
-        RunTswap("init");
-        RunTswap("create", "burn-me");
-        var (exit, stdout, _) = RunTswap("burn", "burn-me", "test reason");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("BURNED", stdout);
-    }
-
-    [Fact]
-    public void Burn_ShowsInNames()
-    {
-        RunTswap("init");
-        RunTswap("create", "visible-burn");
-        RunTswap("burn", "visible-burn", "oops");
-
-        var (_, stdout, _) = RunTswap("names");
-        Assert.Contains("[BURNED]", stdout);
-    }
-
-    [Fact]
-    public void Burn_NonexistentFails()
-    {
-        RunTswap("init");
-        var (exit, _, stderr) = RunTswap("burn", "ghost");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("not found", stderr);
-    }
-
-    [Fact]
-    public void Burn_AlreadyBurnedFails()
-    {
-        RunTswap("init");
-        RunTswap("create", "already-burned");
-        RunTswap("burn", "already-burned", "original reason");
-
-        var (exit, _, stderr) = RunTswap("burn", "already-burned", "second reason");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("already burned", stderr);
-        Assert.Contains("original reason", stderr);
-    }
-
-    // --- Burned ---
-
-    [Fact]
-    public void Burned_ShowsBurnedSecrets()
-    {
-        RunTswap("init");
-        RunTswap("create", "leaked");
-        RunTswap("burn", "leaked", "seen in logs");
-
-        var (exit, stdout, _) = RunTswap("burned");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("leaked", stdout);
-        Assert.Contains("seen in logs", stdout);
-    }
-
-    [Fact]
-    public void Burned_NoBurnedSecrets()
-    {
-        RunTswap("init");
-        RunTswap("create", "clean-secret");
-
-        var (exit, stdout, _) = RunTswap("burned");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("No burned secrets", stdout);
-    }
-
-    // --- Prompt ---
-
-    [Fact]
-    public void Prompt_OutputsInstructions()
-    {
-        // prompt doesn't need init
-        var (exit, stdout, _) = RunTswap("prompt");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("tswap", stdout);
-        Assert.Contains("AI Agent", stdout);
-        // SKILL.md frontmatter
-        Assert.Contains("---", stdout);
-        Assert.Contains("name: tswap", stdout);
-        Assert.Contains("must use when", stdout, StringComparison.OrdinalIgnoreCase);
-    }
-
-    [Fact]
-    public void PromptHash_OutputsHash()
-    {
-        var (exit, stdout, _) = RunTswap("prompt-hash");
-
-        Assert.Equal(0, exit);
-        var hash = stdout.Trim();
-        // SHA-256 hex = 64 chars
-        Assert.Equal(64, hash.Length);
-        Assert.Matches("^[0-9a-f]{64}$", hash);
-    }
-
-    // --- InstallScript ---
-
-    /// <summary>
-    /// Tests invoke the compiled apphost binary directly, so Environment.ProcessPath is
-    /// the binary itself and installscript succeeds and embeds the actual binary path.
-    /// </summary>
-    [Fact]
-    public void InstallScript_OutputsScript()
-    {
-        var (exit, stdout, _) = RunTswap("installscript");
-
-        Assert.Equal(0, exit);
-        if (OperatingSystem.IsWindows())
-        {
-            // PowerShell: no shebang, has ErrorActionPreference
-            Assert.Contains("ErrorActionPreference", stdout);
-            Assert.Contains("tswap.exe", stdout);
-        }
-        else
-        {
-            // Bash: starts with shebang
-            Assert.StartsWith("#!/usr/bin/env bash", stdout.TrimStart());
-            Assert.Contains("tswap", stdout);
-        }
-    }
-
-    [Fact]
-    public void InstallScript_EmbedsBinaryPath()
-    {
-        var (exit, stdout, _) = RunTswap("installscript");
-
-        Assert.Equal(0, exit);
-        // The script must embed a non-empty binary path (not the placeholder)
-        Assert.DoesNotContain("TSWAP_BINARY_PATH_PLACEHOLDER", stdout);
-        // The embedded path must reference tswap
-        Assert.Contains("tswap", stdout);
-    }
-
-    [Fact]
-    public void InstallScript_InstallsSkillMd()
-    {
-        var (exit, stdout, _) = RunTswap("installscript");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("SKILL.md", stdout);
-        Assert.Contains(".agents/skills/tswap", stdout.Replace('\\', '/'));
-    }
-
-    [Fact]
-    public void InstallScript_CreatesClaudeSymlink()
-    {
-        var (exit, stdout, _) = RunTswap("installscript");
-
-        Assert.Equal(0, exit);
-        Assert.Contains(".claude/skills", stdout.Replace('\\', '/'));
-    }
-
-    // --- Run (token substitution) ---
-
-    [Fact]
-    public void Run_SubstitutesToken()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("hello-world", "ingest", "test-val");
-
-        // Use 'true' as the command — it just exits 0, proving substitution worked
-        var (exit, _, _) = RunTswap("run", "true", "{{test-val}}");
-
-        Assert.Equal(0, exit);
-    }
-
-    [Fact]
-    public void Run_MissingTokenFails()
-    {
-        RunTswap("init");
-
-        var (exit, _, stderr) = RunTswap("run", "true", "{{nonexistent}}");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("not found", stderr);
-    }
-
-    [Fact]
-    public void Run_NoTokensFails()
-    {
-        RunTswap("init");
-
-        var (exit, _, stderr) = RunTswap("run", "true");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("No {{tokens}}", stderr);
-    }
-
-    [Fact]
-    public void Run_BlockedCommandFails()
-    {
-        RunTswap("init");
-        RunTswap("create", "block-test");
-
-        var (exit, _, stderr) = RunTswap("run", "echo", "{{block-test}}");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("expose secret", stderr);
-    }
-
-    [Fact]
-    public void Run_BlockedCommandViaTokenSubstitutionFails()
-    {
-        // A token in argv[0] that expands to a blocked command must still be rejected.
-        // The pre-substitution blocklist check sees "{{cmd}}" (not blocked); the
-        // post-substitution re-check must catch the expanded value.
-        RunTswap("init");
-        RunTswapWithStdin("echo", "ingest", "cmd");
-
-        var (exit, _, stderr) = RunTswap("run", "{{cmd}}", "hello");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("expose secret", stderr);
-    }
-
-    [Fact]
-    public void Run_PipeBlockedFails()
-    {
-        RunTswap("init");
-        RunTswap("create", "pipe-test");
-
-        var (exit, _, stderr) = RunTswap("run", "curl", "{{pipe-test}}", "|", "cat");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("Pipes", stderr);
-    }
-
-    [Fact]
-    public void Run_SecretRedactedFromErrorOutput()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("abc123xyz", "ingest", "my-pass");
-
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS())
-            return; // ls + /tmp path is POSIX-only
-
-        // ls on a nonexistent path that includes the secret — ls will echo the path in its
-        // error message. The test harness redirects stdout, so tswap uses FallbackPty, which
-        // merges stderr into stdout when Console.IsOutputRedirected. Redacted text lands on stdout.
-        var (exit, stdout, _) = RunTswap("run", "ls", "/tmp/prefix-{{my-pass}}-suffix");
-
-        Assert.NotEqual(0, exit);
-        Assert.DoesNotContain("abc123xyz", stdout);
-        Assert.Contains("[REDACTED: my-pass]", stdout);
-    }
-
-    [Fact]
-    public void Run_FirstLineOfStdoutNotDropped()
-    {
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS())
-            return; // POSIX sh/echo only
-
-        RunTswap("init");
-        RunTswapWithStdin("s3cr3t", "ingest", "my-secret");
-
-        // Regression test for issue #74 / fix for issue #75.
-        // tswap now execs argv[0] directly (no shell wrapper), so the compound command
-        // is passed as a single literal argument to sh -c — exactly as the user's shell
-        // already parsed it. No artificial quoting workaround is needed.
-        //
-        // Coverage note: the test harness redirects all three streams so tswap always
-        // uses FallbackPty, not UnixPty. This test guards the FallbackPty (pipe) path
-        // and the shared StreamRedactor logic against future regressions.
-        //
-        // 'echo' is not blocked (only the top-level command, "sh", is checked against
-        // the blocklist). The redactor must catch the raw value (s3cr3t) in the output.
-        var (exit, stdout, _) = RunTswap(
-            "run", "sh", "-c",
-            "echo before; echo {{my-secret}}; echo after");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("before",  stdout);
-        Assert.Contains("after",   stdout);
-        Assert.DoesNotContain("s3cr3t", stdout);
-        Assert.Contains("[REDACTED: my-secret]", stdout);
-    }
-
-    /// <summary>
-    /// Regression test for issue #75: compound shell commands passed via sh -c work
-    /// correctly end-to-end. This is the exact user scenario from the bug report:
-    ///   tswap run sh -c 'echo before; echo {{my-secret}}; echo after'
-    /// The user's shell strips the outer single quotes; tswap receives the compound
-    /// command as a single argv element and execs sh directly with it — no intermediate
-    /// bash re-parsing that would split the compound command on semicolons.
-    /// </summary>
-    [Fact]
-    public void Run_CompoundCommandViaShellDashC()
-    {
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS())
-            return; // POSIX sh/echo only
-
-        RunTswap("init");
-        RunTswapWithStdin("s3cr3t", "ingest", "my-secret");
-
-        // Simulate what the user's shell delivers to tswap after stripping outer quotes:
-        // the compound command arrives as one argument, not split on the semicolons.
-        var (exit, stdout, _) = RunTswap(
-            "run", "sh", "-c",
-            "echo before; echo {{my-secret}}; echo after");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("before",  stdout);
-        Assert.Contains("after",   stdout);
-        Assert.DoesNotContain("s3cr3t", stdout);
-        Assert.Contains("[REDACTED: my-secret]", stdout);
-    }
-
-    /// <summary>
-    /// End-to-end test for issues #74 and #75 that exercises the real PTY code path
-    /// (LinuxPty via forkpty). The test opens its own outer PTY and gives tswap the slave
-    /// as its stdin/stdout/stderr via a bash script, so <c>Console.IsOutputRedirected</c>
-    /// returns false inside tswap and <c>PtyRunnerFactory.Create()</c> returns <c>LinuxPty</c> — not
-    /// FallbackPty. Output is captured from the outer PTY master.
-    ///
-    /// tswap now execs argv[0] directly (issue #75 fix), so the compound command string
-    /// is passed as a literal argument to sh -c with no intermediate shell re-parsing.
-    /// </summary>
-    [Fact]
-    public void Run_FirstLineOfStdoutNotDropped_UnixPty()
-    {
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS())
-            return;
-
-        RunTswap("init");
-        RunTswapWithStdin("s3cr3t", "ingest", "my-secret");
-
-        // Open an outer PTY pair. tswap will inherit the slave as its fds so
-        // isatty(0/1/2) returns true and PtyRunnerFactory.Create() picks LinuxPty/MacOSPty.
-        var nameBuf = new byte[256];
-        int masterFd = -1, slaveFd = -1;
-        int ptyRet = OpenPty(ref masterFd, ref slaveFd, nameBuf);
-        if (ptyRet == int.MinValue)
-            return; // openpty not available on this platform; skip test
-        if (ptyRet < 0)
-            throw new Exception($"openpty failed (errno {Marshal.GetLastPInvokeError()})");
-
-        var slavePath = Encoding.ASCII.GetString(nameBuf, 0,
-            Array.IndexOf(nameBuf, (byte)0) is int z and >= 0 ? z : nameBuf.Length);
-
-        // Write a small bash script. The two `exec` lines:
-        //   1. Replace bash's own fds 0/1/2 with the slave PTY (making tswap see a real TTY).
-        //   2. Replace bash with the tswap binary so it inherits those slave fds.
-        //
-        // The compound command is passed as a plain double-quoted string in the bash script.
-        // Bash passes it as a single argv element to tswap, and tswap exec's sh directly
-        // with that element as the -c argument — no intermediate shell re-parsing.
-        var scriptPath = Path.Combine(_tempDir, "pty_test.sh");
-        File.WriteAllText(scriptPath,
-            $"#!/bin/bash\n" +
-            $"exec <\"{slavePath}\" >\"{slavePath}\" 2>\"{slavePath}\"\n" +
-            $"exec \"{_binaryPath}\"" +
-            $" run sh -c \"echo before; echo {{{{my-secret}}}}; echo after\"\n");
-
-        var psi = new ProcessStartInfo { FileName = "bash", UseShellExecute = false, CreateNoWindow = true };
-        psi.ArgumentList.Add(scriptPath);
-        psi.Environment["TSWAP_TEST_KEY"]            = _testKeyHex;
-        psi.Environment["TSWAP_TEST_SUDO_BYPASS"]    = "1";
-        psi.Environment["TSWAP_CONFIG_DIR"]          = _tempDir;
-        psi.Environment["DOTNET_EnableWriteXorExecute"] = "0";
-
-        // Read from the outer PTY master until EIO (all slave fds closed after child exits).
-        // poll() with a 500 ms timeout bounds each read attempt so the test can never hang
-        // indefinitely if tswap or dotnet stalls; the 60 s deadline kills the child process.
-        // try/finally ensures masterFd and slaveFd are closed on all exit paths (assertion
-        // failure, exception from PtyRead, etc.) so the test never leaks file descriptors.
-        const int EINTR  = 4;
-        // EAGAIN differs by OS: Linux=11, macOS=35. Retry rather than treating as EIO
-        // to match the production UnixPty loop and avoid spurious breaks on O_NONBLOCK fds.
-        int eagain = OperatingSystem.IsLinux() ? 11 : 35;
-        const short POLLIN_FLAG = 1;
-        var sb       = new StringBuilder();
-        var buf      = new byte[4096];
-        var deadline = DateTime.UtcNow.AddSeconds(60);
-        string output;
-        using var process = Process.Start(psi)!;
-        try
-        {
-            PtyClose(slaveFd); slaveFd = -1; // close our copy; child holds it
-
-            while (DateTime.UtcNow < deadline)
-            {
-                var pfd = new TestPollFd { fd = masterFd, events = POLLIN_FLAG };
-                int pr = PtyPoll(ref pfd, 1, 500);
-                if (pr == 0) continue; // timeout — check deadline and loop
-                if (pr < 0)
-                {
-                    if (Marshal.GetLastPInvokeError() == EINTR) continue;
-                    break; // poll error
-                }
-                int n = (int)PtyRead(masterFd, buf, (nint)buf.Length);
-                if (n > 0) { sb.Append(Encoding.UTF8.GetString(buf, 0, n)); continue; }
-                if (n == 0) break; // EOF
-                var readErrno = Marshal.GetLastPInvokeError();
-                if (readErrno == EINTR || readErrno == eagain) continue; // interrupted or O_NONBLOCK: retry
-                break; // EIO or other terminal error
-            }
-        }
-        finally
-        {
-            if (masterFd != -1) { PtyClose(masterFd); masterFd = -1; }
-            if (slaveFd  != -1) { PtyClose(slaveFd);  slaveFd  = -1; }
-        }
-        if (!process.WaitForExit(TimeSpan.FromSeconds(5)))
-        {
-            process.Kill(entireProcessTree: true);
-            process.WaitForExit();
-        }
-        output = sb.ToString();
-        Assert.Equal(0, process.ExitCode);
-        Assert.Contains("before",               output);
-        Assert.Contains("after",                output);
-        Assert.DoesNotContain("s3cr3t",         output);
-        Assert.Contains("[REDACTED: my-secret]", output);
-    }
-
-    // --- No args ---
-
-    [Fact]
-    public void NoArgs_ShowsUsage()
-    {
-        var (exit, stdout, _) = RunTswap();
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Usage", stdout);
-    }
-
-    // --- Unknown command ---
-
-    [Fact]
-    public void UnknownCommand_Fails()
-    {
-        var (exit, _, stderr) = RunTswap("bogus");
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("Unknown command", stderr);
-    }
-
-    // --- TSWAP_TEST_KEY validation ---
-
-    [Fact]
-    public void TestKey_WrongLengthFails()
-    {
-        var psi = new ProcessStartInfo
-        {
-            FileName = _binaryPath,
-            RedirectStandardOutput = true,
-            RedirectStandardError = true,
-            UseShellExecute = false,
-            CreateNoWindow = true
-        };
-        psi.ArgumentList.Add("init");
-        psi.Environment["TSWAP_TEST_KEY"] = "AABB"; // Only 2 bytes, not 32
-        psi.Environment["TSWAP_CONFIG_DIR"] = _tempDir;
-
-        using var process = Process.Start(psi)!;
-        var stderr = process.StandardError.ReadToEnd();
-        process.StandardOutput.ReadToEnd();
-        process.WaitForExit();
-
-        Assert.NotEqual(0, process.ExitCode);
-        Assert.Contains("32 bytes", stderr);
-    }
-
-    // --- End-to-end workflow ---
-
-    [Fact]
-    public void EndToEnd_CreateNamesBurnBurned()
-    {
-        // Full lifecycle: init → create → names → burn → burned
-        var (exit1, _, _) = RunTswap("init");
-        Assert.Equal(0, exit1);
-
-        var (exit2, _, _) = RunTswap("create", "e2e-secret");
-        Assert.Equal(0, exit2);
-
-        var (exit3, namesOut, _) = RunTswap("names");
-        Assert.Equal(0, exit3);
-        Assert.Contains("e2e-secret", namesOut);
-        Assert.DoesNotContain("[BURNED]", namesOut);
-
-        var (exit4, _, _) = RunTswap("burn", "e2e-secret", "integration test");
-        Assert.Equal(0, exit4);
-
-        var (exit5, namesOut2, _) = RunTswap("names");
-        Assert.Equal(0, exit5);
-        Assert.Contains("[BURNED]", namesOut2);
-
-        var (exit6, burnedOut, _) = RunTswap("burned");
-        Assert.Equal(0, exit6);
-        Assert.Contains("e2e-secret", burnedOut);
-        Assert.Contains("integration test", burnedOut);
-    }
-
-    [Fact]
-    public void EndToEnd_IngestAndRun()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("test-password-123", "ingest", "my-pass");
-
-        // Use 'test' command to verify the substituted value is non-empty
-        // (can't use redirect/pipe due to exfiltration protection)
-        var (exit, _, _) = RunTswap("run", "test", "-n", "{{my-pass}}");
-
-        Assert.Equal(0, exit);
-    }
-
-    [Fact]
-    public void SameTestKey_ProducesSameMasterKey()
-    {
-        // Verify determinism: same TSWAP_TEST_KEY encrypts/decrypts consistently
-        RunTswap("init");
-        RunTswap("create", "determinism-test");
-
-        // Names should still work (proves decrypt with same key succeeds)
-        var (exit, stdout, _) = RunTswap("names");
-        Assert.Equal(0, exit);
-        Assert.Contains("determinism-test", stdout);
-    }
-
-    // --- Apply ---
-
-    [Fact]
-    public void Apply_OutputsSubstitutedContent()
-    {
-        // Initialize and create a secret
-        RunTswap("init");
-        RunTswap("create", "test-secret", "16");
-
-        // Create a temp YAML file with tswap marker
-        var yamlFile = Path.Combine(_tempDir, "test-values.yaml");
-        File.WriteAllText(yamlFile, @"database:
-  host: localhost
-  password: """"  # tswap: test-secret
-  port: 5432");
-
-        // Run apply and check stdout
-        var (exit, stdout, _) = RunTswap("apply", yamlFile);
-
-        Assert.Equal(0, exit);
-        // Should contain the structure
-        Assert.Contains("database:", stdout);
-        Assert.Contains("host: localhost", stdout);
-        Assert.Contains("port: 5432", stdout);
-        // Should have password line with marker
-        Assert.Contains("# tswap: test-secret", stdout);
-        // Should NOT contain empty password
-        Assert.DoesNotContain(@"password: """"", stdout);
-        // Should have a populated password (16 chars in double quotes)
-        Assert.Matches(@"password: "".{16}""", stdout);
-    }
-
-    [Fact]
-    public void Apply_FailsOnMissingSecret()
-    {
-        RunTswap("init");
-
-        var yamlFile = Path.Combine(_tempDir, "test-values.yaml");
-        File.WriteAllText(yamlFile, @"password: """"  # tswap: nonexistent-secret");
-
-        var (exit, _, stderr) = RunTswap("apply", yamlFile);
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("nonexistent-secret", stderr);
-        Assert.Contains("not found", stderr);
-    }
-
-    // --- Check ---
-
-    [Fact]
-    public void Check_ExitCode1_WhenSecretMissing()
-    {
-        RunTswap("init");
-
-        var yamlFile = Path.Combine(_tempDir, "check-missing.yaml");
-        File.WriteAllText(yamlFile, @"password: """"  # tswap: missing-secret");
-
-        var (exit, _, _) = RunTswap("check", yamlFile);
-
-        Assert.Equal(1, exit);
-    }
-
-    [Fact]
-    public void Check_ExitCode2_WhenSecretBurned()
-    {
-        RunTswap("init");
-        RunTswap("create", "burned-check-secret");
-        RunTswap("burn", "burned-check-secret", "was leaked");
-
-        var yamlFile = Path.Combine(_tempDir, "check-burned.yaml");
-        File.WriteAllText(yamlFile, @"password: """"  # tswap: burned-check-secret");
-
-        var (exit, stdout, _) = RunTswap("check", yamlFile);
-
-        Assert.Equal(2, exit);
-        Assert.Contains("BURNED", stdout);
-    }
-
-    [Fact]
-    public void Check_ExitCode1_WhenMixedMissingAndBurned()
-    {
-        RunTswap("init");
-        RunTswap("create", "burned-mixed-secret");
-        RunTswap("burn", "burned-mixed-secret", "was leaked");
-
-        var yamlFile = Path.Combine(_tempDir, "check-mixed.yaml");
-        File.WriteAllText(yamlFile, @"
-password1: """"  # tswap: burned-mixed-secret
-password2: """"  # tswap: missing-mixed-secret");
-
-        var (exit, stdout, _) = RunTswap("check", yamlFile);
-
-        // Missing takes precedence over burned, so exit code should be 1
-        Assert.Equal(1, exit);
-        Assert.Contains("NOT FOUND", stdout);
-        Assert.Contains("BURNED", stdout);
-    }
-
-    [Fact]
-    public void Check_ExitCode0_WhenAllOk()
-    {
-        RunTswap("init");
-        RunTswap("create", "ok-check-secret");
-
-        var yamlFile = Path.Combine(_tempDir, "check-ok.yaml");
-        File.WriteAllText(yamlFile, @"password: """"  # tswap: ok-check-secret");
-
-        var (exit, _, _) = RunTswap("check", yamlFile);
-
-        Assert.Equal(0, exit);
-    }
-
-    [Fact]
-    public void Apply_DiagnosticsGoToStderr_NotStdout()
-    {
-        // NOTE: This test cannot fully validate the fix without real YubiKey hardware.
-        // In test mode (TSWAP_TEST_KEY set), no security warning banner is generated,
-        // so we can only verify that stdout contains clean YAML without warning text.
-        // The negative assertions pass regardless of whether warnings go to stderr.
-        // 
-        // However, this test still provides value by ensuring that the apply command's
-        // stdout output remains clean YAML that can be safely piped or redirected.
-        
-        RunTswap("init");
-        RunTswap("create", "test-secret", "16");
-
-        var yamlFile = Path.Combine(_tempDir, "test-values.yaml");
-        File.WriteAllText(yamlFile, @"password: """"  # tswap: test-secret");
-
-        var (exit, stdout, stderr) = RunTswap("apply", yamlFile);
-
-        Assert.Equal(0, exit);
-        
-        // Stdout should contain ONLY the YAML output
-        Assert.Contains("password:", stdout);
-        Assert.Contains("# tswap: test-secret", stdout);
-        
-        // Stdout should NOT contain any warning boxes or diagnostic messages
-        // (In test mode these assertions always pass, but they document the requirement)
-        Assert.DoesNotContain("WARNING", stdout);
-        Assert.DoesNotContain("SECURITY", stdout);
-        Assert.DoesNotContain("╔═", stdout);  // Box drawing characters
-        Assert.DoesNotContain("║", stdout);
-        Assert.DoesNotContain("╚═", stdout);
-    }
-
-    // --- Init: new config fields ---
-
-    [Fact]
-    public void Init_ConfigHasUnlockChallenge()
-    {
-        RunTswap("init");
-        var json = File.ReadAllText(Path.Combine(_tempDir, "config.json"));
-        var config = JsonSerializer.Deserialize(json, TswapJsonContext.Default.Config)!;
-
-        Assert.NotNull(config.UnlockChallenge);
-        Assert.Equal(64, config.UnlockChallenge.Length); // 32 bytes as hex
-    }
-
-    [Fact]
-    public void Init_ConfigHasRngMode()
-    {
-        RunTswap("init");
-        var json = File.ReadAllText(Path.Combine(_tempDir, "config.json"));
-        var config = JsonSerializer.Deserialize(json, TswapJsonContext.Default.Config)!;
-
-        Assert.NotNull(config.RngMode);
-        Assert.Equal(RngMode.System, config.RngMode); // test mode uses default
-        // On-disk format compatibility: the enum must serialize as the legacy lowercase string
-        Assert.Contains("\"system\"", json);
-    }
-
-    // Regression test for issue #39: init must not hang when stdin is piped.
-    // The "Insert YubiKey and press Enter" pauses are TTY-only; when stdin is
-    // redirected they must be skipped so that piped "yes" answers only the
-    // reinitialise prompt and the command completes normally.
-    [Fact]
-    public void Init_Reinit_PipedYes_DoesNotHang()
-    {
-        RunTswap("init");
-        // "yes\n" answers the "Already initialized. Reinitialize?" prompt.
-        var (exit, stdout, _) = RunTswapWithStdin("yes\n", "init");
-        Assert.Equal(0, exit);
-        Assert.Contains("test mode", stdout);
-    }
-
-    // --- Export / Import ---
-
-    [Fact]
-    public void Export_CreatesEncryptedFile()
-    {
-        RunTswap("init");
-        RunTswap("create", "my-secret");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        var (exit, stdout, _) = RunTswapWithStdin("passphrase123\npassphrase123\n", "export", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.True(File.Exists(exportPath));
-        // File must not contain the plaintext secret name
-        Assert.DoesNotContain("my-secret", File.ReadAllText(exportPath));
-        Assert.Contains("tswap-export-v1", File.ReadAllText(exportPath));
-    }
-
-    [Fact]
-    public void Export_Import_RoundTrip()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("secret-value", "ingest", "db-pass");
-        RunTswap("create", "api-key");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        RunTswapWithStdin("strongpassphrase\nstrongpassphrase\n", "export", exportPath);
-
-        // Re-init clears the vault (also delete secrets.json.enc — test key is constant
-        // so the old encrypted file remains readable if not explicitly removed)
-        // "yes\n" answers the "Already initialized. Reinitialize?" prompt via stdin.
-        RunTswapWithStdin("yes\n", "init");
-        File.Delete(Path.Combine(_tempDir, "secrets.json.enc"));
-        var (exit, stdout, _) = RunTswapWithStdin("strongpassphrase\n", "import", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Imported 2 secret(s)", stdout);
-
-        var (_, namesOut, _) = RunTswap("names");
-        Assert.Contains("db-pass", namesOut);
-        Assert.Contains("api-key", namesOut);
-    }
-
-    [Fact]
-    public void Import_SkipsBurnedSecrets()
-    {
-        RunTswap("init");
-        RunTswap("create", "good-secret");
-        RunTswap("create", "burned-secret");
-        RunTswap("burn", "burned-secret", "compromised");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        RunTswapWithStdin("passphrase\npassphrase\n", "export", exportPath);
-
-        RunTswapWithStdin("yes\n", "init");
-        File.Delete(Path.Combine(_tempDir, "secrets.json.enc"));
-        var (exit, stdout, _) = RunTswapWithStdin("passphrase\n", "import", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Skipped", stdout);
-        Assert.Contains("burned-secret", stdout);
-
-        var (_, namesOut, _) = RunTswap("names");
-        Assert.Contains("good-secret", namesOut);
-        Assert.DoesNotContain("burned-secret", namesOut);
-    }
-
-    [Fact]
-    public void Import_SkipsExistingSecrets()
-    {
-        RunTswap("init");
-        RunTswap("create", "existing-secret");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        RunTswapWithStdin("passphrase\npassphrase\n", "export", exportPath);
-
-        // Import into the same vault — existing-secret already there
-        var (exit, stdout, _) = RunTswapWithStdin("passphrase\n", "import", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Skipped", stdout);
-        Assert.Contains("existing-secret", stdout);
-    }
-
-    [Fact]
-    public void Import_WrongPassphraseFails()
-    {
-        RunTswap("init");
-        RunTswap("create", "a-secret");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        RunTswapWithStdin("correct-passphrase\ncorrect-passphrase\n", "export", exportPath);
-
-        var (exit, _, stderr) = RunTswapWithStdin("wrong-passphrase\n", "import", exportPath);
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("wrong passphrase", stderr);
-    }
-
-    [Fact]
-    public void Import_SkipsSecretsWithNulOrNullValues()
-    {
-        RunTswap("init");
-
-        // Construct a synthetic export file containing one clean secret, one with a NUL byte,
-        // and one with a null value (possible from a tampered/corrupted export file since
-        // System.Text.Json can produce null for non-nullable string properties).
-        // We bypass tswap for this because ingest now rejects NUL values.
-        const string passphrase = "test-passphrase";
-        var salt       = RandomNumberGenerator.GetBytes(16);
-        var exportKey  = Crypto.DeriveKeyFromPassphrase(passphrase, salt);
-
-        // Inject the null-value secret via raw JSON to bypass the C# type system.
-        var dbJson = $$"""
-            {
-              "Secrets": {
-                "good-secret": {"Value":"good-value","Created":"{{DateTime.UtcNow:O}}","Modified":"{{DateTime.UtcNow:O}}"},
-                "nul-secret":  {"Value":"bad\u0000val","Created":"{{DateTime.UtcNow:O}}","Modified":"{{DateTime.UtcNow:O}}"},
-                "null-secret": {"Value":null,          "Created":"{{DateTime.UtcNow:O}}","Modified":"{{DateTime.UtcNow:O}}"}
-              }
-            }
-            """;
-        var ciphertext = Crypto.Encrypt(Encoding.UTF8.GetBytes(dbJson), exportKey);
-        var exportFile = new ExportFile(
-            "tswap-export-v1", DateTime.UtcNow,
-            Convert.ToBase64String(salt),
-            Convert.ToBase64String(ciphertext));
-        var exportPath = Path.Combine(_tempDir, "nul-test.enc");
-        File.WriteAllText(exportPath,
-            JsonSerializer.Serialize(exportFile, TswapJsonContext.Default.ExportFile));
-
-        var (exit, stdout, _) = RunTswapWithStdin(passphrase + "\n", "import", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Imported 1 secret(s)", stdout);
-        Assert.Contains("nul-secret",  stdout);
-        Assert.Contains("null-secret", stdout);
-
-        var (_, namesOut, _) = RunTswap("names");
-        Assert.Contains("good-secret",      namesOut);
-        Assert.DoesNotContain("nul-secret",  namesOut);
-        Assert.DoesNotContain("null-secret", namesOut);
-    }
-
-    // --- Migrate ---
-
-    [Fact]
-    public void Migrate_OldConfig_UpdatesRngMode()
-    {
-        // Simulate a pre-feature config: init normally then strip UnlockChallenge
-        RunTswap("init");
-        var configPath = Path.Combine(_tempDir, "config.json");
-        var config = JsonSerializer.Deserialize(
-            File.ReadAllText(configPath), TswapJsonContext.Default.Config)!;
-        var oldConfig = config with { UnlockChallenge = null, RngMode = null };
-        File.WriteAllText(configPath,
-            JsonSerializer.Serialize(oldConfig, TswapJsonContext.Default.Config));
-
-        // Run migrate, choose YubiKey entropy mode (option 2), decline detailed instructions
-        var (exit, stdout, _) = RunTswapWithStdin("2\nno\n", "migrate");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("YubiKey hardware", stdout);
-
-        var updated = JsonSerializer.Deserialize(
-            File.ReadAllText(configPath), TswapJsonContext.Default.Config)!;
-        Assert.Equal(RngMode.YubiKey, updated.RngMode);
-    }
-
-    [Fact]
-    public void Migrate_ModernConfig_ReportsUpToDate()
-    {
-        RunTswap("init");
-        // Modern config has UnlockChallenge set and RequiresTouch from synthetic init
-        // We need to also set RequiresTouch = true to satisfy all checks
-        var configPath = Path.Combine(_tempDir, "config.json");
-        var config = JsonSerializer.Deserialize(
-            File.ReadAllText(configPath), TswapJsonContext.Default.Config)!;
-        var modernConfig = config with { RequiresTouch = true };
-        File.WriteAllText(configPath,
-            JsonSerializer.Serialize(modernConfig, TswapJsonContext.Default.Config));
-
-        var (exit, stdout, _) = RunTswap("migrate");
-
-        Assert.Equal(0, exit);
-        Assert.Contains("up to date", stdout);
-    }
-
-    // Regression test for issue #39: migrate must not hang when stdin is piped.
-    // Synthetic init leaves RequiresTouch=null → needsTouchMigration=true → the
-    // "Show detailed re-initialization instructions? (yes/no):" prompt is shown.
-    // Piped "no\n" must answer it without blocking and the command must exit 0.
-    [Fact]
-    public void Migrate_PipedNo_DoesNotHang()
-    {
-        RunTswap("init");
-        // Default synthetic config: RngMode="system", UnlockChallenge=set, RequiresTouch=null
-        // → needsReInit=true (touch migration required) → yes/no prompt is shown
-        var (exit, _, _) = RunTswapWithStdin("no\n", "migrate");
-        Assert.Equal(0, exit);
-    }
-
-    // --- tocomment: security ---
-
-    [Fact]
-    public void ToComment_DryRun_DoesNotExposeSecretInOutput()
-    {
-        RunTswap("init");
-        // Ingest a known secret value via stdin
-        var (ingestExit, _, _) = RunTswapWithStdin("supersecretvalue123", "ingest", "my-password");
-        Assert.Equal(0, ingestExit);
-
-        // Write a YAML file containing the plaintext secret
-        var yamlFile = Path.Combine(_tempDir, "applied.yaml");
-        File.WriteAllText(yamlFile, "password: supersecretvalue123");
-
-        // Run tocomment --dry-run and capture output
-        var (exit, stdout, stderr) = RunTswap("tocomment", yamlFile, "--dry-run");
-
-        Assert.Equal(0, exit);
-        // The secret must NOT appear anywhere in the output (stdout or stderr)
-        Assert.DoesNotContain("supersecretvalue123", stdout);
-        Assert.DoesNotContain("supersecretvalue123", stderr);
-        // The diff output is now in stderr (diagnostic output)
-        Assert.Contains("[REDACTED: my-password]", stderr);
-        Assert.Contains("# tswap: my-password", stderr);
-    }
-
-    [Fact]
-    public void ToComment_Live_DoesNotExposeSecretInOutput()
-    {
-        RunTswap("init");
-        var (ingestExit, _, _) = RunTswapWithStdin("supersecretvalue123", "ingest", "my-password");
-        Assert.Equal(0, ingestExit);
-
-        var yamlFile = Path.Combine(_tempDir, "applied.yaml");
-        File.WriteAllText(yamlFile, "password: supersecretvalue123");
-
-        var (exit, stdout, stderr) = RunTswap("tocomment", yamlFile);
-
-        Assert.Equal(0, exit);
-        // Secret must not appear in either stream
-        Assert.DoesNotContain("supersecretvalue123", stdout);
-        Assert.DoesNotContain("supersecretvalue123", stderr);
-        // Diff output is in stderr
-        Assert.Contains("[REDACTED: my-password]", stderr);
-    }
-
-    // --- PTY P/Invoke helpers (used by Run_FirstLineOfStdoutNotDropped_UnixPty) ---
-
-    // openpty: try libc first (glibc 2.34+ / Fedora 35+ / Ubuntu 22.04+), then libutil.
-    [DllImport("libc",    EntryPoint = "openpty", SetLastError = true)]
-    private static extern int openpty_libc(ref int amaster, ref int aslave, [Out] byte[] name, IntPtr termp, IntPtr winp);
-
-    [DllImport("libutil", EntryPoint = "openpty", SetLastError = true)]
-    private static extern int openpty_libutil(ref int amaster, ref int aslave, [Out] byte[] name, IntPtr termp, IntPtr winp);
-
-    [DllImport("libc", EntryPoint = "read",  SetLastError = true)]
-    private static extern nint PtyRead(int fd, [Out] byte[] buf, nint count);
-
-    [DllImport("libc", EntryPoint = "write", SetLastError = true)]
-    private static extern nint PtyWrite(int fd, [In] byte[] buf, nint count);
-
-    [DllImport("libc", EntryPoint = "close")]
-    private static extern int PtyClose(int fd);
-
-    [StructLayout(LayoutKind.Sequential)]
-    private struct TestPollFd { public int fd; public short events; public short revents; }
-
-    [DllImport("libc", EntryPoint = "poll", SetLastError = true)]
-    private static extern int PtyPoll(ref TestPollFd fds, uint nfds, int timeout);
-
-    // Returns the openpty result (< 0 on failure), or int.MinValue if openpty is not
-    // available on this platform (neither libc nor libutil exports the symbol).
-    private static int OpenPty(ref int masterFd, ref int slaveFd, byte[] nameBuf)
-    {
-        try { return openpty_libc(ref masterFd, ref slaveFd, nameBuf, IntPtr.Zero, IntPtr.Zero); }
-        catch (DllNotFoundException) { }
-        catch (EntryPointNotFoundException) { } // openpty not exported by libc on this platform
-        try { return openpty_libutil(ref masterFd, ref slaveFd, nameBuf, IntPtr.Zero, IntPtr.Zero); }
-        catch (DllNotFoundException) { }
-        catch (EntryPointNotFoundException) { } // libutil not present or openpty not exported
-        return int.MinValue; // sentinel: openpty not available on this platform
-    }
-
-    // Starts tswap with a PTY so ReadPassword enters interactive mode, then sends
-    // Ctrl+C (0x03) until the process exits.  Two triggers fire a send (subject to a
-    // 1-second cooldown between sends):
-    //   • quiescent: 500 ms of silence after any output — process is blocked in ReadKey
-    //   • prompt match: waitForPrompt text appears in accumulated output
-    // Early sends (before TreatControlCAsInput=true is set) generate SIGINT with no
-    // foreground pgid, which is silently dropped; the process keeps running and more
-    // output arrives, resetting the quiescent timer for another attempt.
-    // Returns (exitCode, output); (-2, "") if openpty is unavailable on this platform.
-    private (int exitCode, string output) RunTswapWithPtyCtrlC(string waitForPrompt, params string[] args)
-    {
-        var nameBuf = new byte[256];
-        int masterFd = -1, slaveFd = -1;
-        int ptyRet = OpenPty(ref masterFd, ref slaveFd, nameBuf);
-        if (ptyRet == int.MinValue) return (-2, "");
-        if (ptyRet < 0) throw new Exception($"openpty failed (errno {Marshal.GetLastPInvokeError()})");
-
-        var slavePath = Encoding.ASCII.GetString(nameBuf, 0,
-            Array.IndexOf(nameBuf, (byte)0) is int z and >= 0 ? z : nameBuf.Length);
-
-        var quotedArgs = string.Join(" ", args.Select(a => $"\"{a}\""));
-        var scriptPath = Path.Combine(_tempDir, $"pty_ctrlc_{Guid.NewGuid():N}.sh");
-        File.WriteAllText(scriptPath,
-            $"#!/bin/bash\n" +
-            $"exec <\"{slavePath}\" >\"{slavePath}\" 2>\"{slavePath}\"\n" +
-            $"exec \"{_binaryPath}\" {quotedArgs}\n");
-
-        var psi = new ProcessStartInfo { FileName = "bash", UseShellExecute = false, CreateNoWindow = true };
-        psi.ArgumentList.Add(scriptPath);
-        psi.Environment["TSWAP_TEST_KEY"] = _testKeyHex;
-        psi.Environment["TSWAP_TEST_SUDO_BYPASS"] = "1";
-        psi.Environment["TSWAP_CONFIG_DIR"] = _tempDir;
-        psi.Environment["DOTNET_EnableWriteXorExecute"] = "0";
-
-        var sb = new StringBuilder();
-        var buf = new byte[4096];
-        var deadline = DateTime.UtcNow.AddSeconds(120);
-        DateTime? lastDataAt = null;
-        DateTime lastCtrlCAt = DateTime.MinValue;
-        bool promptSeen = false;
-        // searchFrom tracks how much of sb has already been scanned for waitForPrompt,
-        // leaving a (prompt.Length - 1) overlap to catch matches that straddle chunks.
-        int searchFrom = 0;
-        const int EINTR = 4;
-        int eagain = OperatingSystem.IsLinux() ? 11 : 35;
-
-        void TrySendCtrlC()
-        {
-            if (DateTime.UtcNow - lastCtrlCAt < TimeSpan.FromMilliseconds(1000)) return;
-            // Retry on EINTR/EAGAIN so Ctrl+C is always delivered.
-            var ctrlC = new byte[] { 0x03 };
-            nint written;
-            do { written = PtyWrite(masterFd, ctrlC, 1); }
-            while (written < 0 && Marshal.GetLastPInvokeError() is int e && (e == EINTR || e == eagain));
-            if (written == 1) lastCtrlCAt = DateTime.UtcNow;
-        }
-
-        // Open the try/finally before Process.Start so fd cleanup runs even if Start throws.
-        try
-        {
-            using var process = Process.Start(psi)!;
-            PtyClose(slaveFd); slaveFd = -1;
-
-            while (DateTime.UtcNow < deadline)
-            {
-                var pfd = new TestPollFd { fd = masterFd, events = 1 }; // POLLIN
-                int pr = PtyPoll(ref pfd, 1, 200);
-                if (pr == 0)
-                {
-                    // No new data for 200 ms. Fire if quiescent for 500 ms total.
-                    if (lastDataAt.HasValue &&
-                        DateTime.UtcNow - lastDataAt.Value > TimeSpan.FromMilliseconds(500))
-                        TrySendCtrlC();
-                    continue;
-                }
-                if (pr < 0)
-                {
-                    if (Marshal.GetLastPInvokeError() == EINTR) continue;
-                    break;
-                }
-                int n = (int)PtyRead(masterFd, buf, (nint)buf.Length);
-                if (n > 0)
-                {
-                    sb.Append(Encoding.UTF8.GetString(buf, 0, n));
-                    lastDataAt = DateTime.UtcNow;
-                    // Scan only the newly appended region (plus overlap) rather than
-                    // re-materialising the full buffer on every chunk.
-                    if (!promptSeen)
-                    {
-                        int scanFrom = Math.Max(0, searchFrom - waitForPrompt.Length + 1);
-                        if (sb.ToString(scanFrom, sb.Length - scanFrom).Contains(waitForPrompt))
-                            promptSeen = true;
-                        searchFrom = sb.Length;
-                    }
-                    if (promptSeen)
-                        TrySendCtrlC();
-                    continue;
-                }
-                if (n == 0) break; // EOF
-                var errno = Marshal.GetLastPInvokeError();
-                if (errno == EINTR || errno == eagain) continue;
-                break; // EIO or other terminal error
-            }
-
-            if (!process.WaitForExit(TimeSpan.FromSeconds(10)))
-            {
-                process.Kill(entireProcessTree: true);
-                process.WaitForExit();
-            }
-
-            return (process.ExitCode, sb.ToString());
-        }
-        finally
-        {
-            if (masterFd != -1) { PtyClose(masterFd); masterFd = -1; }
-            if (slaveFd  != -1) { PtyClose(slaveFd);  slaveFd  = -1; }
-        }
-    }
-
-    [Fact]
-    public void ToComment_ContinuationLine_DoesNotLeakFragmentInOutput()
-    {
-        RunTswap("init");
-        var (ingestExit, _, _) = RunTswapWithStdin("supersecretvalue123", "ingest", "my-password");
-        Assert.Equal(0, ingestExit);
-
-        // Simulates a multi-line YAML scalar: the continuation line is a base64-looking
-        // fragment. RedactContent cannot redact it (it is only a partial match), so without
-        // the fix it would be printed verbatim — the test ensures it is suppressed instead.
-        var yamlFile = Path.Combine(_tempDir, "multiline.yaml");
-        File.WriteAllText(yamlFile,
-            "password: supersecretvalue123\n" +
-            "  AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA\n");
-
-        var (exit, stdout, stderr) = RunTswap("tocomment", yamlFile, "--dry-run");
-
-        Assert.Equal(0, exit);
-        // Raw continuation fragment must NOT appear in either stream
-        Assert.DoesNotContain("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA", stdout);
-        Assert.DoesNotContain("AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA", stderr);
-        // Safe placeholder must appear in stderr (diff output)
-        Assert.Contains("[removed continuation line]", stderr);
-        // Main secret must also not be leaked in either stream
-        Assert.DoesNotContain("supersecretvalue123", stdout);
-        Assert.DoesNotContain("supersecretvalue123", stderr);
-    }
-
-    // --- issue #69: init creates vault file ---
-
-    [Fact]
-    public void Init_CreatesVaultFile()
-    {
-        var (exit, _, _) = RunTswap("init");
-
-        Assert.Equal(0, exit);
-        Assert.True(File.Exists(Path.Combine(_tempDir, "secrets.json.enc")),
-            "init should create an empty secrets.json.enc so subsequent commands do not warn about a missing vault");
-    }
-
-    [Fact]
-    public void Names_WarnsMissingVaultOnStderr()
-    {
-        RunTswap("init");
-        File.Delete(Path.Combine(_tempDir, "secrets.json.enc"));
-
-        var (exit, _, stderr) = RunTswap("names");
-
-        // Command succeeds (empty vault) and emits a warning on stderr
-        Assert.Equal(0, exit);
-        Assert.Contains("vault file not found", stderr, StringComparison.OrdinalIgnoreCase);
-    }
-
-    // --- issue #67: export prompts go to stderr ---
-
-    [Fact]
-    public void Export_PromptsAreOnStderr()
-    {
-        RunTswap("init");
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        var (exit, stdout, stderr) = RunTswapWithStdin("testpass\ntestpass\n", "export", exportPath);
-
-        Assert.Equal(0, exit);
-        // The interactive prompt "Export passphrase:" must not pollute stdout
-        // (stdout is the redirect target when the user pipes export output).
-        Assert.DoesNotContain("Export passphrase:", stdout);
-        Assert.DoesNotContain("Confirm passphrase:", stdout);
-        // Prompts must appear on stderr so the user sees them when stdout is redirected
-        Assert.Contains("Export passphrase:", stderr);
-        Assert.Contains("Confirm passphrase:", stderr);
-    }
-
-    // --- issue #68: export cancel exits with non-zero code ---
-
-    [Fact]
-    public void Export_Cancel_ExitsWithError()
-    {
-        RunTswap("init");
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        File.WriteAllText(exportPath, "existing"); // trigger the overwrite prompt
-
-        // Answer "no" to the overwrite question — should cancel and exit non-zero
-        var (exit, _, stderr) = RunTswapWithStdin("no\n", "export", exportPath);
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("cancelled", stderr, StringComparison.OrdinalIgnoreCase);
-        // File must not be modified
-        Assert.Equal("existing", File.ReadAllText(exportPath));
-    }
-
-    // --- issue #29: import with invalid JSON gives friendly error ---
-
-    [Fact]
-    public void Import_InvalidJson_GivesFriendlyError()
-    {
-        RunTswap("init");
-        var badFile = Path.Combine(_tempDir, "bad.enc");
-        File.WriteAllText(badFile, "this is not json at all {{{");
-
-        var (exit, _, stderr) = RunTswapWithStdin("anypassphrase\n", "import", badFile);
-
-        Assert.NotEqual(0, exit);
-        Assert.Contains("not valid JSON", stderr, StringComparison.OrdinalIgnoreCase);
-    }
-
-    // --- issue #81: extend burn handling ---
-
-    [Fact]
-    public void Redact_IncludesBurnedSecretValues()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("burned-secret-value", "ingest", "my-burned");
-        RunTswap("burn", "my-burned", "test");
-
-        var yamlFile = Path.Combine(_tempDir, "content.txt");
-        File.WriteAllText(yamlFile, "token: burned-secret-value");
-
-        var (exit, stdout, _) = RunTswap("redact", yamlFile);
-
-        Assert.Equal(0, exit);
-        Assert.DoesNotContain("burned-secret-value", stdout);
-        Assert.Contains("[REDACTED: my-burned]", stdout);
-    }
-
-    [Fact]
-    public void ToComment_IncludesBurnedSecretValues()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("burned-secret-value", "ingest", "my-burned");
-        RunTswap("burn", "my-burned", "test");
-
-        var yamlFile = Path.Combine(_tempDir, "content.yaml");
-        File.WriteAllText(yamlFile, "token: burned-secret-value");
-
-        var (exit, stdout, stderr) = RunTswap("tocomment", yamlFile, "--dry-run");
-
-        Assert.Equal(0, exit);
-        Assert.DoesNotContain("burned-secret-value", stdout);
-        Assert.DoesNotContain("burned-secret-value", stderr);
-        Assert.Contains("# tswap: my-burned", stderr);
-    }
-
-    [Fact]
-    public void Apply_BurnedSecret_AppliesWithWarning()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("burned-secret-value", "ingest", "my-burned");
-        RunTswap("burn", "my-burned", "test");
-
-        var yamlFile = Path.Combine(_tempDir, "apply-burned.yaml");
-        File.WriteAllText(yamlFile, "token: \"\"  # tswap: my-burned");
-
-        var (exit, stdout, stderr) = RunTswap("apply", yamlFile);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("burned-secret-value", stdout);
-        Assert.Contains("Warning", stderr);
-        Assert.Contains("burned", stderr);
-    }
-
-    [Fact]
-    public void Import_IncludeBurned_ImportsBurnedSecrets()
-    {
-        RunTswap("init");
-        RunTswapWithStdin("burned-value", "ingest", "was-burned");
-        RunTswap("burn", "was-burned", "compromised");
-        RunTswap("create", "good-secret");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        RunTswapWithStdin("passphrase\npassphrase\n", "export", exportPath);
-
-        RunTswapWithStdin("yes\n", "init");
-        File.Delete(Path.Combine(_tempDir, "secrets.json.enc"));
-        var (exit, stdout, _) = RunTswapWithStdin("passphrase\n", "import", "--include-burned", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Imported 2 secret(s)", stdout);
-
-        var (_, namesOut, _) = RunTswap("names");
-        Assert.Contains("good-secret", namesOut);
-        Assert.Contains("was-burned", namesOut);
-        Assert.Contains("[BURNED]", namesOut);
-    }
-
-    [Fact]
-    public void Import_Default_SkipsBurnedSecrets_WithHint()
-    {
-        RunTswap("init");
-        RunTswap("create", "good-secret");
-        RunTswap("create", "burned-secret");
-        RunTswap("burn", "burned-secret", "compromised");
-
-        var exportPath = Path.Combine(_tempDir, "backup.enc");
-        RunTswapWithStdin("passphrase\npassphrase\n", "export", exportPath);
-
-        RunTswapWithStdin("yes\n", "init");
-        File.Delete(Path.Combine(_tempDir, "secrets.json.enc"));
-        var (exit, stdout, _) = RunTswapWithStdin("passphrase\n", "import", exportPath);
-
-        Assert.Equal(0, exit);
-        Assert.Contains("Skipped", stdout);
-        Assert.Contains("burned-secret", stdout);
-        Assert.Contains("--include-burned", stdout);
-
-        var (_, namesOut, _) = RunTswap("names");
-        Assert.Contains("good-secret", namesOut);
-        Assert.DoesNotContain("burned-secret", namesOut);
-    }
-
-    // --- issue #85: Ctrl+C during ReadPassword restores terminal and prints "Cancelled" ---
-    // These tests require a PTY so ReadPassword takes the interactive path.
-    // Before the fix, Ctrl+C raised SIGINT and the process died without printing "Cancelled".
-    // After the fix, TreatControlCAsInput intercepts it, the finally block resets the flag,
-    // OperationCanceledException propagates to the top-level handler, and "Cancelled" is printed.
-
-    [Fact]
-    public void Add_CtrlC_DuringPasswordPrompt_PrintsCancelledAndExits130()
-    {
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS()) return;
-
-        RunTswap("init");
-
-        var (exitCode, output) = RunTswapWithPtyCtrlC("Secret value for", "add", "ctrlc-secret");
-
-        if (exitCode == -2) return; // openpty not available on this platform
-
-        Assert.True(exitCode == 130, $"Expected exit 130, got {exitCode}. PTY output: [{output}]");
-        Assert.True(output.Contains("Cancelled"), $"Expected 'Cancelled' in PTY output: [{output}]");
-    }
-
-    [Fact]
-    public void Export_CtrlC_DuringPassphrasePrompt_PrintsCancelledAndExits130()
-    {
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS()) return;
-
-        RunTswap("init");
-        RunTswap("create", "a-secret");
-        var exportPath = Path.Combine(_tempDir, "ctrlc-export.enc");
-
-        var (exitCode, output) = RunTswapWithPtyCtrlC("Export passphrase:", "export", exportPath);
-
-        if (exitCode == -2) return; // openpty not available on this platform
-
-        Assert.True(exitCode == 130, $"Expected exit 130, got {exitCode}. PTY output: [{output}]");
-        Assert.True(output.Contains("Cancelled"), $"Expected 'Cancelled' in PTY output: [{output}]");
-        Assert.False(File.Exists(exportPath), "export file must not be created on cancellation");
-    }
-
-    [Fact]
-    public void Import_CtrlC_DuringPassphrasePrompt_PrintsCancelledAndExits130()
-    {
-        if (!OperatingSystem.IsLinux() && !OperatingSystem.IsMacOS()) return;
-
-        RunTswap("init");
-        RunTswap("create", "a-secret");
-        var exportPath = Path.Combine(_tempDir, "ctrlc-import.enc");
-        RunTswapWithStdin("passphrase\npassphrase\n", "export", exportPath);
-
-        var (exitCode, output) = RunTswapWithPtyCtrlC("Import passphrase:", "import", exportPath);
-
-        if (exitCode == -2) return; // openpty not available on this platform
-
-        Assert.True(exitCode == 130, $"Expected exit 130, got {exitCode}. PTY output: [{output}]");
-        Assert.True(output.Contains("Cancelled"), $"Expected 'Cancelled' in PTY output: [{output}]");
-    }
-}
