+++ TswapTests/TswapBinaryFixture.cs
@@ -18,6 +18,17 @@ public sealed class TswapBinaryFixture
 
     public TswapBinaryFixture()
     {
+        // CI (and local runs) can point the E2E suite at a pre-built binary — e.g. a
+        // NativeAOT publish — instead of building the Debug apphost here.
+        var overridePath = Environment.GetEnvironmentVariable("TSWAP_E2E_BINARY");
+        if (!string.IsNullOrEmpty(overridePath))
+        {
+            BinaryPath = Path.GetFullPath(overridePath);
+            if (!File.Exists(BinaryPath))
+                throw new Exception($"TSWAP_E2E_BINARY points to a missing file: {BinaryPath}");
+            return;
+        }
+
         var projectDir = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "..", "..", "..", ".."));
         var psi = new ProcessStartInfo
         {
