+++ TswapTests/TswapTests.csproj
@@ -6,6 +6,9 @@
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
+    <!-- TswapCli is a RID-specific (self-contained-on-publish) exe; referencing it from
+         this RID-agnostic test project is fine for in-process command tests. -->
+    <ValidateExecutableReferencesMatchSelfContained>false</ValidateExecutableReferencesMatchSelfContained>
   </PropertyGroup>
 
   <ItemGroup>
@@ -16,6 +19,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\TswapCore\TswapCore.csproj" />
+    <ProjectReference Include="..\TswapCli\TswapCli.csproj" />
   </ItemGroup>
 
   <ItemGroup>
